% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_board_from_lcaname.R
\name{add_board_from_lcaname}
\alias{add_board_from_lcaname}
\title{Assign Health Board name based on Local Authority name}
\usage{
add_board_from_lcaname(df, lcaname, region = FALSE)
}
\arguments{
\item{df}{The dataframe you want to mutate}

\item{lcaname}{The name of the LCA variable}

\item{region}{If FALSE, returns Health Board in "NHS XXXXXX". If TRUE, returns "XXXXXX Region"}
}
\value{
Dataframe with an extra column named hb
}
\description{
This function should pick up all common variations in LCA name, such as 'City of Edinburgh'
vs 'Edinburgh City'. Any non-matching LCA names will return 'Other Non-Scottish Residents'
}
\examples{
lcaname_to_hb(bluenile::lca_names, lca, region = FALSE)
}
\seealso{
Other Recoding: 
\code{\link{add_hospital_board}()}
}
\concept{Recoding}
