% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hospital_board.R
\name{add_hospital_board}
\alias{add_hospital_board}
\title{Assign Health Board names based on Hospital Location codes}
\usage{
add_hospital_board(df, locationcode, hbcode, region = FALSE)
}
\arguments{
\item{df}{The data frame to be mutated}

\item{locationcode}{The name of the location code variable}

\item{hbcode}{The name of the existing Health Board code variable}

\item{region}{When TRUE, returns "XXXX Region", when FALSE returns "NHS XXXX" (default FALSE)}
}
\value{
Data frame
}
\description{
Assign Health Board names based on Hospital Location codes
}
\examples{
example_data <- base::data.frame(location = c("A0001", "B0002", "C0004"), hbres = c("", "", "S08000007"))
example_data_2 <- add_hospital_board(example_data, location, hbres, region = TRUE)
}
\seealso{
Other Recoding: 
\code{\link{add_board_from_lcaname}()}
}
\concept{Recoding}
