% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_slf.R
\name{extract_from_slf}
\alias{extract_from_slf}
\title{Extract the correct variables for Source workbooks from the Source Linkage Files}
\usage{
extract_from_slf(year, workbook_name, sub_name)
}
\arguments{
\item{year}{The year or years to be extracted, in "XXYY" format}

\item{workbook_name}{A valid workbook name from \link{workbook_names}}

\item{sub_name}{For HRI Pathways and HRI Main, specification of which part is being worked on}
}
\description{
A wrapper around \link[slfhelper:read_slf_individual]{slfhelper::read_slf_individual} and \link[slfhelper:read_slf_episode]{slfhelper::read_slf_episode} that
defines the columns to be extracted based on pre-defined list within bluenile
}
\examples{
x <- utils::head(extract_from_slf(year = "1819", workbook_name = "HRI", sub_name = "main"), 1000)
}
