% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_the_slf.R
\name{filter_the_slf}
\alias{filter_the_slf}
\title{Filter the SLF Extract based on the workbook data being created}
\usage{
filter_the_slf(df, workbook_name, level)
}
\arguments{
\item{workbook_name}{See workbook_names for valid list}

\item{level}{The security level of the workbook. Valid options are formatted "levelx"}
}
\value{
Filtered dataframe
}
\description{
Filter the SLF Extract based on the workbook data being created
}
\examples{
workbook_name <- "LTC"
slfhelper::read_slf_individual("1819", from = 10000, to = 100000) \%>\% filter_the_slf(workbook_name, "level3")
}
