% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrf_dir.R
\name{get_nrf_dir}
\alias{get_nrf_dir}
\title{Source lookup directory - National Reference Files}
\usage{
get_nrf_dir()
}
\value{
The path to the NRFs as an \link[fs:path]{fs::path}
}
\description{
File path for the general PHS directory for National References Files
}
\concept{Directories}
