% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specialty_lookup.R
\name{get_specialty_lookup}
\alias{get_specialty_lookup}
\title{Read in the Specialty Grouping lookup file from National Reference Files}
\usage{
get_specialty_lookup(var_grouping = FALSE)
}
\arguments{
\item{var_grouping}{When FALSE, removes the variable 'grouping' (default FALSE)}
}
\value{
A data frame containing specialty codes and the names of the specialties
}
\description{
Read in the Specialty Grouping lookup file from National Reference Files
}
