% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hea_constants}
\alias{hea_constants}
\title{List of vectors, relating to variables in the production of
Hospital-Expenditure-and-Activity}
\format{
An object of class \code{list} of length 6.
}
\usage{
hea_constants
}
\description{
List of vectors, relating to variables in the production of
Hospital-Expenditure-and-Activity
}
\keyword{data}
