% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lca_names}
\alias{lca_names}
\title{A tibble of LCA names and 9-digit codes (CA2019)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 2 columns.
}
\usage{
lca_names
}
\description{
For use in examples or as a lookup. Names match those from phsmethods::match_area()
}
\keyword{data}
