% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locality_lookup.R, R/lookup_functions.R
\name{locality_lookup}
\alias{locality_lookup}
\title{Read in most recent Geography team locality lookup file}
\usage{
locality_lookup(suffix, columns)

locality_lookup(suffix, columns)
}
\arguments{
\item{suffix}{A string in "YYYYMMDD" format correlating to most recent geography file}

\item{columns}{A character vector of columns}
}
\value{
A tibble with the selected columns to use as a lookup

A tibble with the selected columns to use as a lookup
}
\description{
Read in most recent Geography team locality lookup file

Read in most recent Geography team locality lookup file
}
\examples{
x <- locality_lookup("20200825", c("datazone2011", "hscp_locality"))

x <- locality_lookup("20200825", c("datazone2011", "hscp_locality"))

}
