% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locality_lookup.R, R/lookup_functions.R
\name{specialty_lookup}
\alias{specialty_lookup}
\title{Read in a lookup file for Specialty codes}
\usage{
specialty_lookup(columns)

specialty_lookup(columns)
}
\arguments{
\item{columns}{A vector of columns to pass to dplyr::select.
Available columns are "speccode", "description", "grouping"}
}
\value{
A tibble to use as a lookup for specialty groupings

A tibble to use as a lookup for specialty groupings
}
\description{
Read in a lookup file for Specialty codes

Read in a lookup file for Specialty codes
}
\examples{
x <- specialty_lookup(c("speccode", "description"))

x <- specialty_lookup(c("speccode", "description"))

}
