% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{twodigitla_to_hb}
\alias{twodigitla_to_hb}
\title{Assign 9-digit Health Board code based on two-digit LA code}
\usage{
twodigitla_to_hb(df, lcacodename)
}
\arguments{
\item{df}{The dataframe you wish to mutate}

\item{lcacodename}{The name of the two-digit LA code variable. Can be formatted as integer or character}
}
\value{
Mutated dataframe, with extra column 'hbres'
}
\description{
Assign 9-digit Health Board code based on two-digit LA code
}
\examples{
example_data <- base::data.frame(la = c(1, 3, 5, 7))
twodigitla_to_hb(example_data, la)

example_data_2 <- base::data.frame(la = c("01", "03", "05", "13"))
twodigitla_to_hb(example_data_2, la)
}
