% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_to_extract}
\alias{variables_to_extract}
\title{List of vectors, each member of which is a variable from the SLFs.}
\format{
An object of class \code{list} of length 7.
}
\usage{
variables_to_extract
}
\description{
Used to extract correct variables from SLF
}
\keyword{data}
