% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{where_is_hospital}
\alias{where_is_hospital}
\title{Function to assign Health Board names based on 5-digit hospital location code}
\usage{
where_is_hospital(df, locationcode, hbcode)
}
\arguments{
\item{df}{The data frame you want to change}

\item{locationcode}{The name of the location code variable in df}

\item{hbcode}{The name of the Health Board code in df}
}
\value{
df with one extra column, hosp_board
}
\description{
Function to assign Health Board names based on 5-digit hospital location code
}
\examples{
example_data <- base::data.frame(location = c("A0001", "B0002", "C0004"), hbres = c("", "", "S08000007"))
where_is_hospital(example_data, location, hbres)
}
