% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{workbook_names}
\alias{workbook_names}
\title{List with three levels, one for each security level on Source platform}
\format{
An object of class \code{list} of length 3.
}
\usage{
workbook_names
}
\description{
For reference
}
\keyword{data}
